Implement TCP communication function
This commit adds an incomplete TCP communication function to the codebase. The function tcp_communication is responsible for handling incoming TCP connections, receiving data from the client, and sending it back. It uses the accept, recv, and send functions from the sys/socket.h header file. Note that this function is not fully implemented and needs further work.